Simple: Do not add extra fake nodes for ECMP
The base reasoning was that since fake nodes
change the perceived metric of the routes,
in order to preserve existing ECMP present
before augmenting the topology, we had to
back-propagate these change in order to
maintain that ECMP cost. An easy solution
was thus to set up fake nodes on all ECMP
nodes, regardless of changes in their
nexthop sets.

While true for Merger, this makes no sense
for simple as the metric change is only seen
locally. As such, we no longer do this for
Simple.

Signed-off-by: Olivier Tilmans <olivier.tilmans@uclouvain.be>